Just did a quick test of the match page and found some errors in the JavaScript.
When it was contacting /advisors/magic_match_test_page.php, the JavaScript wasn't sending
the session_id properly which caused the match algorithm to not run. It was due to a spelling mistake
on the variable which has been fixed.

I also added a small blurb to the README and will add more later.