I am currently working on the pagination of the search results. To make life easier: I split the match page into two pages: match.php and html/results_section_match_page.html.
The latter page contains the html jsut for the results section. It is inserted into the page using PHP's file_get_contents()
function.

I created a PHP file called get_result_page.php that resturns a "page" of results given $page, $limit, $type GET variables.
It works and has been tested.

The HTML is acting very odd and won't place buttons at the bottom of the results section. For now, I will place buttons to
the left (underneath the delimiters) just to test the pagination. I will ask for Konstantin's help in with the HTML.