The new version of the explore page is complete. It's currently in explore2.html but will soon replace explore.html (which was used for demoing purposes).
In explore.html:

The search bar was made into a directive since it's used on multiple pages.
Eventually, all of the pages will use the search bar directive instead of
building it from scratch each time.

Began working on importing CSV files into the database. It appears to work, but
couldn't test it since the file I was given did not contain all of the
pertinent information.

It appears that to get the auto scraper working with the google spreadsheet, I
will have to rewrite a lot of the code. It was designed to work in a very
specific way and is not very flexible. Now that I know the format of the
information, I should be able to build a more robust crawler.