Moved explore2.html to explore.html
Reworked the Import.IO scraper to work with Fordham: successfully inserted
advisors, course, and funding (no theses provided). It hasn't been tested for
the other universities, but it should still work since most of the code is the
same, just moved around to be more efficient.

TODO: Full test of scraper, include CSV and PDF scraping