Implemented a new funding delimiter that allows users to view only advisors that have funding associated with them.
The signup/registration overlay was extended to all pages in the website. To keep the code clean, the overlay was put into its own view and
service. The view isn't run with Angular but instead is PHP. This can be changed later into a directive for ease of use and cleaner code.

Fixed the edit profile glitch so users can now update their profile.

TODO: Implement thesis delimiter where only advisors that have overseen student theses & projects are displayed.